cleanup text logger
